Add pytorch

+++ src/_numpy.py
@@ -1,4 +1,5 @@
 import numpy as np
 
 def rootloop(N):
-	return np.sum( np.sqrt( np.arange(0, N, 2, dtype=int) ) )
+	# Use float to be comparable with pyTorch
+	return np.sum( np.sqrt( np.arange(0.0, N, 2) ) )
